-Implement OMAP address lookup support -Implement support for detecting UNWIND_INFO.ExceptionData based on various heuristics -Implement FuncInfo4

+++ src/PESpy/Native/03. Exception/FuncInfoHeader.cs
@@ -19,6 +19,6 @@
             };
     #pragma warning(pop)
         };*/
-        public int value;
+        public byte value;
     }
 }
